HW#9 Object Mapper and DTO

+++ src/main/java/com/example/restshops/service/impl/ShopServiceImpl.java
@@ -4,6 +4,8 @@ import com.example.restshops.repository.models.Shop;
 import com.example.restshops.repository.ShopRepo;
 import com.example.restshops.service.ShopService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -18,19 +20,32 @@ public class ShopServiceImpl implements ShopService {
     }
 
     @Override
-    public Optional<Shop> getById(Long id) {
-        if (id != null)
+    public ResponseEntity<?> getById(Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } else {
+            return new ResponseEntity<>(shopRepo.findById(id),HttpStatus.FOUND);
+        }
+    }
+
+    @Override
+    public Optional<Shop> getByIdObj(Long id) {
+        if (id != null) {
             return shopRepo.findById(id);
-        else return Optional.empty();
+        } else return Optional.empty();
     }
 
     @Override
-    public Shop save(Shop shop) {
-        return shopRepo.save(shop);
+    public ResponseEntity<?> save(Shop shop) {
+        if (shop == null) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(shopRepo.save(shop),HttpStatus.CREATED);
+        }
     }
 
     @Override
-    public Optional<Shop> update(Shop shop) {
+    public ResponseEntity<?> update(Shop shop) {
         Optional<Shop> shopOptional = shopRepo.findById(shop.getId());
         if (shopOptional.isPresent()) {
             Shop foundedShop = shopOptional.get();
@@ -39,14 +54,19 @@ public class ShopServiceImpl implements ShopService {
             foundedShop.setTitle(shop.getTitle());
             foundedShop.setNumberOfWorkers(shop.getNumberOfWorkers());
             foundedShop.setSite(shop.isSite());
-            Shop updatedShop = shopRepo.save(foundedShop);
-            return Optional.of(updatedShop);
+            return new ResponseEntity<>(shopRepo.save(foundedShop),HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return Optional.empty();
     }
 
     @Override
-    public void deleteShop(Long id) {
-        shopRepo.deleteById(id);
+    public ResponseEntity<?> deleteShop(Long id) {
+        if (shopRepo.findById(id).isPresent()) {
+            shopRepo.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
